pyproject.toml
[project]
name = "rest-incendio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==4.2.11",
    "djangorestframework==3.14.0",
    "numpy>=1.26.0",
    "requests>=2.33.1",
    "scipy>=1.12.0",
]

[tool.setuptools]
packages = ["config", "simulation"]
